refactor: split new space modal
