+++ admin-ui/src/component/Unit/SpacesResources.tsx
@@ -18,7 +18,7 @@ import ResourcesTable from "./ResourcesTable";
 import SpacesTable from "./SpacesTable";
 import SubPageHead from "./SubPageHead";
 import Modal, { useModal as useHDSModal } from "../HDSModal";
-import NewSpaceModal from "../Spaces/space-editor/NewSpaceModal";
+import NewSpaceModal from "../Spaces/space-editor/new-space-modal";
 import { breakpoints } from "../../styles/util";
 import NewResourceModal from "./NewResourceModal";
 import { UNIT_QUERY } from "../../common/queries";
