+++ admin-ui/src/component/Unit/SpacesTable.tsx
@@ -9,7 +9,7 @@ import { DELETE_SPACE } from "../../common/queries";
 import DataTable, { CellConfig } from "../DataTable";
 import PopupMenu from "./PopupMenu";
 import Modal, { useModal as useHDSModal } from "../HDSModal";
-import NewSpaceModal from "../Spaces/space-editor/NewSpaceModal";
+import NewSpaceModal from "../Spaces/space-editor/new-space-modal";
 import ConfirmationDialog, { ModalRef } from "../ConfirmationDialog";
 
 import {
